Populate beatmap set `difficulty_names` after queued difficulty calculation completes
Notably this currently only does something in the normal "difficulty
calculation queued via redis" case. The commands don't attempt to run
this logic. The reason is that it's going to be very annoying to make
that work because the commands are based on beatmap IDs and not beatmap
set IDs, and I don't know that it's worth the time investment to have
this be 100% absolutely correct even after a rework or a manual fix via
the commands.

I'm basically putting this out because I was prepared to do work to fix
it but lost confidence midway through, and there are possible five other
ways of fixing this issue that may be preferable at this stage.